requirements.txt
langchain-core
langchain-groq
python-dotenv
streamlit
python-docx
PyPDF2
faiss-cpu
sentence-transformers
langchain
langchain_community